visitNode as a separate module
